Fixes numpy type error

+++ EyeTrackApp/one_euro_filter.py
@@ -20,7 +20,7 @@ class OneEuroFilter:
         self.beta = np.full(x0.shape, beta)
         self.d_cutoff = np.full(x0.shape, d_cutoff)
         # Previous values.
-        self.x_prev = x0.astype(np.float)
+        self.x_prev = x0
         self.dx_prev = np.full(x0.shape, dx0)
         self.t_prev = time()
